+++ runtime/protocol/http-client/common/src/aws/smithy/kotlin/runtime/http/operation/OperationRequest.kt
@@ -5,7 +5,7 @@
 
 package aws.smithy.kotlin.runtime.http.operation
 
-import aws.smithy.kotlin.runtime.client.ExecutionContext
+import aws.smithy.kotlin.runtime.operation.ExecutionContext
 import aws.smithy.kotlin.runtime.util.CanDeepCopy
 
 /**
