+++ runtime/protocol/http-client/common/test/aws/smithy/kotlin/runtime/http/operation/TracingTestInterceptor.kt
@@ -13,6 +13,7 @@ import aws.smithy.kotlin.runtime.http.interceptors.HttpInterceptor
 import aws.smithy.kotlin.runtime.http.request.HttpRequest
 import aws.smithy.kotlin.runtime.http.response.HttpCall
 import aws.smithy.kotlin.runtime.http.response.HttpResponse
+import aws.smithy.kotlin.runtime.operation.ExecutionContext
 import aws.smithy.kotlin.runtime.time.Instant
 
 class TestException(override val message: String?) : IllegalStateException()
