+++ runtime/smithy-client/common/src/aws/smithy/kotlin/runtime/client/ClientOptionsBuilder.kt
@@ -5,6 +5,7 @@
 
 package aws.smithy.kotlin.runtime.client
 
+import aws.smithy.kotlin.runtime.operation.ExecutionContext
 import aws.smithy.kotlin.runtime.util.AttributeKey
 import aws.smithy.kotlin.runtime.util.Attributes
 import kotlin.properties.ReadWriteProperty
